backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pandas
playwright
requests
pytz
yfinance
feedparser
newspaper3k
lxml
readability-lxml
google.genai
asyncio
nest_asyncio